Refactor Recap to focus on schemas
I've just done yet-another-major-refactor (tm) to recenter Recap around
schemas.

Previously, Recap had a fairly flexible crawler and function registry.
Developers could add arbitrary Pydantic data types to the function registry.
Recap also had the half-baked concept of lineage and data access. Lastly, the
catalog could be transparently accessed via the CLI (like a read-through cache).

I've eliminated all of that. Things are much simpler. Recap now has the
following components:

1. A `type` module that contains Recap types (int, struct, and so on). Think of
   this as Recap's schema AST.
2. A `models` module that contains a very basic Recap `types.Type` wrapper
   implemented as a Pydantic model. Think of `schema.Schema` as the data model
   that's described in the spec. It gets parsed into the AST.
3. `converters`, which convert between different IDLs (Avro, Protobuf, Recap,
   and so on).
4. `diff`, which diffs Recap types.
6. A `reader` module, which lets developers fetch schemas from remote locations
   (using `fsspec` and `sqlalchemy`).
5. A `catalog`, which includes a REST client and server and a very basic
   (single table) database storage layer. The catalog also includes a crawler,
   which uses readers to read schemas and store write them to the catalog REST API.

Putting all of this together, you can do things like diff schemas:

```python
import recap

schema_1 = recap.schema("postgresql://user:pass@localhost/my_db/public/table1")
schema_2 = recap.schema("postgresql://user:pass@localhost/my_db/public/table2")

recap.diff(schema_1, schema_2)
```

Diff'ing works across schema formats, too:

```python
import recap

schema_1 = recap.schema("postgresql://user:pass@localhost/my_db/public/table1")
schema_2 = recap.schema("s3://my_bucket/table1.json")

recap.diff(schema_1, schema_2)
```

You can also transpile schemas:

```python
import recap

jsc = {
    "type": "object",
    "title": "AddressBook",
    "properties": {
        "number": {"type": "number"},
        "street_name": {"type": "string"},
    },
    "required": ["number", "street_name"],
}
avro_schema = recap.convert(jsc, from_format="json", to_format="avro")
```

And populate the catalog with schemas from different data sources:

```bash
$ recap serve
$ recap crawl snowflake://user:pass@host/MY_DB/public?warehouse=COMPUTE_WH
$ recap crawl gs://my_bucket
```

Now, you might be wondering if lineage, data profiling, and data access are
coming back. The answer is: it depends. I am open to it if I find it fits well
with use cases that users have. But I've learned (from my deep-dive into
schemas) that getting modeling right is non-trivial, so I would want to spend
significant time getting the model and features right. That time is being spent
on schemas now, so I'm removing the half-assed stuff I had.
