--- recap/integrations/__init__.py
@@ -1,9 +1,0 @@
-# Load all integrations into the registry
-
-try:
-    import recap.integrations.bigquery
-except:
-    # Skip Google if dependencies aren't installed.
-    pass
-import recap.integrations.fsspec
-import recap.integrations.sqlalchemy
