--- recap/schema/converters/__init__.py
@@ -1,34 +1,0 @@
-from recap.schema.converters.converter import Converter
-
-Converter.register_converter(
-    "avro",
-    "recap.schema.converters.avro.AvroConverter",
-)
-Converter.register_converter(
-    "bigquery",
-    "recap.schema.converters.avro.BigQueryConverter",
-)
-Converter.register_converter(
-    "frictionless",
-    "recap.schema.converters.avro.FrictionlessConverter",
-)
-Converter.register_converter(
-    "protobuf",
-    "recap.schema.converters.avro.ProtobufConverter",
-)
-Converter.register_converter(
-    "sqlalchemy",
-    "recap.schema.converters.avro.SQLAlchemyConverter",
-)
-Converter.register_converter(
-    "sqlglot",
-    "recap.schema.converters.sqlglot.SQLGlotConverter",
-)
-Converter.register_converter(
-    "json",
-    "recap.schema.converters.json_schema.JsonSchemaConverter",
-)
-Converter.register_converter(
-    "recap",
-    "recap.schema.converters.recap.RecapConverter",
-)
